move ratelimit into workload section, move generators into workload
