+++ configs/momento.toml
@@ -43,14 +43,14 @@ poolsize = 1
 concurrency = 20
 # the connect timeout in milliseconds
 connect_timeout = 10000
-# set a global ratelimit for requests
-request_ratelimit = 50
 # set the timeout in milliseconds
 request_timeout = 1000
 
 [workload]
-# the number of threads that will be used to generate requests
+# the number of threads that will be used to generate the workload
 threads = 1
+# set a global ratelimit for the workload
+ratelimit = 50
 
 # An example keyspace showcasing the use of the `key-value` family of commands.
 #
