+++ configs/ping.toml
@@ -39,14 +39,14 @@ threads = 4
 poolsize = 20
 # the connect timeout in milliseconds
 connect_timeout = 10000
-# set a global ratelimit for requests
-request_ratelimit = 10_000
 # set the timeout in milliseconds
 request_timeout = 1000
 
 [workload]
-# the number of threads that will be used to generate requests
+# the number of threads that will be used to generate the workload
 threads = 1
+# set a global ratelimit for the workload
+ratelimit = 10_000
 
 # Note, even though the command does not use keys, it's still a member of a
 # keyspace.
