+++ src/config/client.rs
@@ -14,9 +14,6 @@ pub struct Client {
     connect_timeout: u64,
     /// Request timeout
     request_timeout: u64,
-    // zero is treated as unlimited
-    #[serde(default)]
-    request_ratelimit: u64,
     // number of threads for client tasks
     threads: usize,
     /// Specifies the rate at which connections should randomly reconnect. This
@@ -36,10 +33,6 @@ impl Client {
 
     pub fn request_timeout(&self) -> Duration {
         Duration::from_millis(self.request_timeout)
-    }
-
-    pub fn request_ratelimit(&self) -> Option<NonZeroU64> {
-        NonZeroU64::new(self.request_ratelimit)
     }
 
     pub fn poolsize(&self) -> usize {
